update task 16

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -7,13 +7,13 @@
  */
 int main(void)
 {
-	unsigned long long i, x = 1, y = 2, z;
+	unsigned long i, x = 1, y = 2, z;
 
-	printf("%llu, %llu, ", x, y);
+	printf("%lu, %lu, ", x, y);
 	for (i = 1; i <= 96; i++)
 	{
 		z = x + y;
-		printf(", %llu", z);
+		printf(", %lu", z);
 		x = y;
 		y = z;
 	}
